Create valiations folder
